Move all nexus file functions into common module
