+++ main.py
@@ -8,7 +8,7 @@ from PySide2.QtGui import QIcon
 from PySide2.QtWidgets import QApplication, QMainWindow
 from PySide2 import QtCore
 from nexus_constructor.main_window import MainWindow
-from nexus_constructor.nexus_wrapper import NexusWrapper
+from nexus_constructor.nexus import NexusWrapper
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
