+++ nexus_constructor/nexus/__init__.py
@@ -1,0 +1,6 @@
+from .nexus_wrapper import *
+from .geometry import *
+
+"""
+Functions for manipulating the data in the in-memory NeXus file
+"""
