+++ tests/ui_tests/test_ui_add_component_window.py
@@ -1,7 +1,7 @@
 from PySide2.QtWidgets import QMainWindow, QDialog
 from nexus_constructor.add_component_window import AddComponentDialog
 from nexus_constructor.main_window import MainWindow
-from nexus_constructor.nexus_wrapper import NexusWrapper
+from nexus_constructor.nexus import NexusWrapper
 from PySide2.QtCore import Qt
